// cintia/sin-ia/angular/restaurantes/src/app/interfaces/restaurante.ts
export interface Restaurante {
    id: number;
    nombre: string;
    descripcion: string;
    fotografia_url: string;
    latitud: number;
    longitud: number;
    total_recetas?: number;
    rating_summary?: {
        average: number;
        count: number;
        distribution: Record<string, number>;
    }
}